requirements.txt
apache-beam[gcp]==2.53.0
pandas==2.2.0
google-cloud-bigquery==3.17.1
google-cloud-storage==2.14.0
python-dateutil==2.8.2